chore: 구조 변화
변화
